% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plantecophys-package.R
\docType{data}
\name{acidata1}
\alias{acidata1}
\title{An example A-Ci curve}
\format{\describe{
\item{CO2S}{CO2 concentration in cuvette (ppm)}
\item{Ci}{Intercellular CO2 concentration (ppm)}
\item{Tleaf}{Leaf temperature (deg C)}
\item{Photo}{Net photosynthesis rate (mu mol m-2 s-1)}
}}
\description{
CO2 response of leaf photosynthesis, as measured with a Licor6400.
}

