% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findCiTransition.R
\name{findCiTransition}
\alias{findCiTransition}
\title{Calculate transition points for fitted A-Ci curves}
\usage{
findCiTransition(object, ...)
}
\arguments{
\item{object}{Either an object returned by \code{fitaci}, or a copy of the \code{Photosyn} function.}

\item{\dots}{Further arguments passed to the \code{Photosyn} function.}
}
\description{
Calculates the Ci at the transition points between Ac & Aj (point 1), and Aj and Ap (point 2). The latter is not NA only when TPU was estimated (and estimable), see \code{\link{fitaci}}, argument \code{fitTPU}.
}
\details{
This function is also used by \code{fitaci}, the results are stored in elements \code{Ci_transition} and \code{Ci_transition2}.
}
