% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fitBB.R
\name{fitBB}
\alias{fitBB}
\title{Fit Ball-Berry type models of stomatal conductance}
\usage{
fitBB(df, varnames = list(ALEAF = "Photo", GS = "Cond", VPD = "VpdL", Ca =
  "CO2S", RH = "RH"), gsmodel = c("BBOpti", "BBLeuning", "BallBerry"),
  fitg0 = FALSE)
}
\arguments{
\item{df}{Input dataframe, containing all variables needed to fit the model.}

\item{varnames}{List of names of variables in the input dataframe. Relative humidity (RH) is only
needed when the original Ball-Berry model is to be fit.}

\item{gsmodel}{One of BBOpti (Medlyn et al. 2011), BBLeuning (Leuning 1995), or BallBerry (Ball and Berry 1987).}

\item{fitg0}{If TRUE, also fits the intercept term (g0, the 'residual conductance'). Default is FALSE.}
}
\description{
Fits one of three versions of the Ball-Berry type stomatal conductance models to
observations of stomatal conductance (gs), photosynthesis (A), atmospheric CO2 concentration (Ca)
and vapour pressure deficit (VPD).
}
\references{
#' Leuning, R. 1995. A critical-appraisal of a combined stomatal-photosynthesis model for C-3 plants. Plant Cell and Environment. 18:339-355.

Medlyn, B.E., R.A. Duursma, D. Eamus, D.S. Ellsworth, I.C. Prentice, C.V.M. Barton, K.Y. Crous, P. De Angelis, M. Freeman and L. Wingate. 2011. Reconciling the optimal and empirical approaches to modelling stomatal conductance. Global Change Biology. 17:2134-2144.
}

