% Generated by roxygen2 (4.0.1): do not edit by hand
\name{gamplot}
\alias{gamplot}
\title{Plot a simple generalized additive model fit}
\usage{
gamplot(xvarname, yvarname, groupname = "", dfr, add = FALSE, CI = TRUE,
  CIlevel = 0.95, polycolor = "grey", line = TRUE, linecolor = "black",
  lwd = 2, optline = FALSE, gamfit = NULL, ...)
}
\arguments{
\item{xvarname}{Name of the X variable (quoted)}

\item{yvarname}{Name of the Y variable (quoted)}

\item{groupname}{Name of the grouping variable (quoted, optional)}

\item{dfr}{Dataframe that contains the above three variables}

\item{add}{Logical, if TRUE adds to current plot}

\item{CI}{Logical, if TRUE (the default), adds a shaded confidence region}

\item{CIlevel}{The coverage of the confidence region (defaults to 0.95).}

\item{polycolor}{Colour of the polygon}

\item{line}{Logical, if TRUE adds a line for the mean prediction}

\item{linecolor}{Color of the line (can be a vector with colors for each group)}

\item{lwd}{Width of the line}

\item{optline}{Logical, if TRUE adds a vertical line at the maximum value of Y.}

\item{gamfit}{Optional, an object returned by \code{gam}, which can yield predictions for the dfr provided as input.}
}
\value{
The fitted gam object, invisibly.
}
\description{
A wrapper for \code{gam} from the \code{mgcv} package, to produce a  smoother with a confidence region.
}
\details{
If a fitted gam object is not provided, this function fits a vanilla smoother function when there is no grouping variable,
\preformatted{
gamfit <- gam(Y ~ s(X), data=dat)
}

When a grouping variable is provided, it fits :
\preformatted{
gamfit <- gam(Y ~ group + s(X, by=group, id=1), data=dat)
}

See the documentation for \code{\link{mgcv::gam}} in the \code{mgcv} package.
}

