% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plantecophys-package.R
\docType{data}
\name{manyacidat}
\alias{manyacidat}
\title{An example dataset with multiple A-Ci curves}
\format{
\describe{
\item{Curve}{An identifier for the A-Ci curve (28 curves in total, 13-14 points per curve)}
\item{Ci}{Intercellular CO2 concentration (ppm)}
\item{Photo}{Net photosynthesis rate (mu mol m-2 s-1)}
\item{Tleaf}{Leaf temperature (deg C)}
\item{PPFD}{Photosynthetic photon flux density (mu mol m-2 s-1)}
}
}
\description{
CO2 response of leaf photosynthesis, as measured with a Licor6400, for multiple leaves.
}
