% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitacis.R
\name{fitacis}
\alias{fitacis}
\title{Fit multiple A-Ci curves at once}
\usage{
fitacis(data, group, fitmethod = c("default", "bilinear"),
  progressbar = TRUE, quiet = FALSE, ...)
}
\arguments{
\item{data}{Dataframe with Ci, Photo, Tleaf, PPFD (the last two are optional). For \code{fitacis}, also requires a grouping variable.}

\item{group}{The name of the grouping variable in the dataframe (an A-Ci curve will be fit for each group separately).}

\item{progressbar}{Display a progress bar (default is TRUE).}

\item{quiet}{If TRUE, no messages are written to the screen.}

\item{\dots}{Further arguments passed to \code{\link{fitaci}}}
}
\description{
Fit multiple A-Ci curves at once
}

