% Generated by roxygen2 (4.0.2): do not edit by hand
\name{PhotosynEB}
\alias{FindTleaf}
\alias{PhotosynEB}
\title{Coupled leaf gas exchange model with energy balance}
\usage{
PhotosynEB(Tair = 25, VPD = 1.5, Wind = 2, Wleaf = 0.02,
  StomatalRatio = 1, LeafAbs = 0.86, RH = NULL, ...)

FindTleaf(gs, Tair, ...)
}
\arguments{
\item{Tair}{Air temperature (C)}

\item{VPD}{The vapour pressure deficit of the air (i.e. not the leaf-to-air VPD) (kPa).}

\item{Wind}{Wind speed (m s-1)}

\item{Wleaf}{Leaf width (m)}

\item{StomatalRatio}{The stomatal ratio (cf. Licor6400 terminology), if it is 1, leaves have stomata only on one side (hypostomatous), 2 for leaves with stomata on both sides (amphistomatous).}

\item{LeafAbs}{Leaf absorptance of solar radiation (0-1).}

\item{RH}{The relative humidity of the air (i.e. not calculated with leaf temperature) (in percent).}

\item{gs}{For \code{FindTleaf}, the stomatal conductance (mol m-2 s-1).}

\item{\dots}{Further parameters passed to \code{\link{Photosyn}}. Note that Tleaf is not allowed as an input, since that is calculated by \code{PhotosynEB} from energy balance.}
}
\description{
As \code{\link{Photosyn}}, but calculates the leaf temperature based on the leaf's energy balance. Including sensible and long-wave heat loss, latent heat loss from evaporation, and solar radiation input.

Calculate Tleaf from energy balance, given that we know gs
}
\details{
Uses the Penman-Monteith equation to calculate the leaf transpiration rate, and finds Tleaf by solving the leaf energy balance iteratively. In the solution, it is accounted for that stomatal conductance and net radiation depend on Tleaf. There are no simplifications or approximations made to find the solution.

Also included is the function \code{FindTleaf}, which calculates the leaf temperature if the stomatal conductance is known.
}

